% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.triangle.R
\name{is.triangle}
\alias{is.triangle}
\title{is.triangle}
\usage{
is.triangle(x, quiet = FALSE)
}
\arguments{
\item{x}{The object top be tested.}

\item{quiet}{Logical, default is FALSE. I TRUE will print a messages if it's not a valid triangle.}
}
\value{
Logical
}
\description{
Test whether the supplied object is a valid trangle.
}
\details{
Returns TRUE if:

- x is a numeric vector
- Has length 3
- The values are all positive
- The values define a genometrically valid triangle of volume > 0 (i.e. it's not a line)

Returns FALSE otherwise
}
\examples{

is.triangle(c(3,4,5))

}
