% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.triangle.R
\name{print.triangle}
\alias{print.triangle}
\title{Print a triangle object}
\usage{
\method{print}{triangle}(x)
}
\arguments{
\item{x}{The triangle}
}
\description{
Print a triangle object
}
\examples{

t <- triangle(c(3,4,5))
print(t)

}
