% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{standard_triangle_mle}
\alias{standard_triangle_mle}
\title{Maximum likelihood estimate of the standard triangle distribution mode}
\usage{
standard_triangle_mle(x, debug = FALSE)
}
\arguments{
\item{x}{sample from a triangle distribution}

\item{debug}{if \code{TRUE} then the function will check the input parameters}
}
\value{
an object of S3 class \code{triangle_mle} containing a list with the call, coefficients,
variance co-variance matrix, minimum negative log likelihood,
number of observations, and the sample
}
\description{
Maximum likelihood estimate of the standard triangle distribution mode
}
\examples{
xtest <- c(0.1, 0.25, 0.3, 0.4, 0.45, 0.6, 0.75, 0.8)
standard_triangle_mle(xtest)
}
\references{
\url{https://www.worldscientific.com/doi/suppl/10.1142/5720/suppl_file/5720_chap1.pdf}
}
