% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.triangle.R
\name{summary.triangle}
\alias{summary.triangle}
\title{Print summary details of a triangle object}
\usage{
\method{summary}{triangle}(x)
}
\arguments{
\item{x}{The triangle}
}
\description{
Print summary details of a triangle object
}
\examples{

t <- triangle(c(3,4,5))
summary(t)

}
