% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_function.R
\name{triangle}
\alias{triangle}
\title{Pythagorean theorem}
\usage{
triangle(a, b)
}
\arguments{
\item{a}{= Side one of right triangle, b = Side two of right triangle}
}
\description{
This function allows for the use of the Pythhagorean theorem, a^2+b^2=c^2
}
\examples{
triangle()
}
\keyword{Pythagroean,}
\keyword{triangle}
