% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.R
\name{triangle_mom}
\alias{triangle_mom}
\title{Triangle distribution method of moments estimate}
\usage{
triangle_mom(x, na.rm = FALSE, type = 1)
}
\arguments{
\item{x}{triangle distribution sample}

\item{na.rm}{whether to remove NA samples.  Default is FALSE.}

\item{type}{the type of method of moments.  Type 1 uses the min and max.  Type2
minimizes the distance from the calculated mean, variance, and skewness to the
sample mean, variance, and skewness}
}
\value{
a named vector of the parameter estimates
}
\description{
Triangle distribution method of moments estimate
}
\examples{
set.seed(1204)
x <- rtriangle(20, 0, 2, 1.5)
triangle_mom(x)
}
