% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-plots.R
\name{compare_triangle_fit}
\alias{compare_triangle_fit}
\title{Compare multiple triangle distributions fits}
\usage{
compare_triangle_fit(
  y,
  cols = c("red", "blue", "green"),
  main = "Triangle Fit Comparison",
  ...
)
}
\arguments{
\item{y}{the triangle distributed sample}

\item{cols}{the colors of the CDF-based estimates, the maximum likelihood estimates, and the method of moments estimates}

\item{main}{the plot title}

\item{...}{other parameters passed to \code{plot.ecdf}}
}
\description{
Compare multiple triangle distributions fits
}
\examples{
xtest <- c(0.1, 0.25, 0.3, 0.4, 0.45, 0.6, 0.75, 0.8)
compare_triangle_fit(xtest)
}
