% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{triangle}
\alias{triangle}
\title{triangle}
\usage{
triangle(x)
}
\arguments{
\item{x}{Vector of triangle side lengths}
}
\value{
Triangle object
}
\description{
return a triangle object from the supplied vector
}
\examples{

t <- triangle(c(3,4,5))

}
