% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfe.R
\name{triangle_cdfe}
\alias{triangle_cdfe}
\title{Triangle parameter estimates using a non-linear fit of the empirical CDF}
\usage{
triangle_cdfe(x)
}
\arguments{
\item{x}{the triangle distributed sample}
}
\value{
an object of class \code{nls}
}
\description{
Triangle parameter estimates using a non-linear fit of the empirical CDF
}
\examples{
xtest <- c(0.1, 0.25, 0.3, 0.4, 0.45, 0.6, 0.75, 0.8)
cdfe <- triangle_cdfe(xtest)
print(cdfe)
summary(cdfe)
coef(cdfe)
if (isNamespaceLoaded("MASS")) confint(cdfe)
}
