% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.R
\name{triangle_mom}
\alias{triangle_mom}
\title{Triangle distribution method of moments estimate}
\usage{
triangle_mom(x)
}
\arguments{
\item{x}{triangle distribution sample}
}
\value{
a named vector of the parameter estimates
}
\description{
Triangle distribution method of moments estimate
}
\examples{
set.seed(1204)
x <- rtriangle(20, 0, 2, 1.5)
triangle_mom(x)
}
